evaluation/requirements.txt
absl-py==2.1.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
bert-score==0.3.13
cached-property==1.5.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cycler==0.11.0
datasets==2.13.2
dill==0.3.6
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.12.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.10
importlib-metadata==6.7.0
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.5.3
multidict==6.0.5
multiprocess==0.70.15
nltk==3.8.1
numpy==1.21.6
openai==1.39.0
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
pyarrow==12.0.1
pydantic==2.5.3
pydantic-core==2.14.6
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.2
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
rouge-score==0.1.2
safetensors==0.4.5
six==1.16.0
sniffio==1.3.1
tokenizers==0.13.3
torch==1.13.1
tqdm==4.66.5
transformers==4.30.2
typing-extensions==4.7.1
urllib3==2.0.7
xxhash==3.5.0
yarl==1.9.4
zipp==3.15.0